implemented signup as a component
